+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Zie de blogpost op https://www.dso-toolkit.nl voor meer informatie over de break
 
 ### Tasks
 * Packages: Dependencies updates ([#2718](https://github.com/dso-toolkit/dso-toolkit/issues/2718))
+* Packages: `ESLint` updaten naar v9 ([#2631](https://github.com/dso-toolkit/dso-toolkit/issues/2631))
 
 ## 🐘 Release 62.30.1 - 2024-07-29